Implemented Authentication to Journal Endpoints

+++ src/main/java/com/spring/JournalApplication/service/JournalEntryService.java
@@ -34,7 +34,7 @@ public class JournalEntryService {
         }
     }
 
-    public void createJournalEntry(JournalEntry journalEntry) {
+    public void updateJournalEntry(JournalEntry journalEntry) {
         journalEntry.setDateTime(LocalDateTime.now());
         journalEntryRepository.save(journalEntry);
     }
@@ -47,10 +47,18 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
+    @Transactional
     public void deleteJournalEntryById(ObjectId id, String userName) {
-        User user = userService.findByUserName(userName);
-        journalEntryRepository.deleteById(id);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveUser(user);
+        try {
+            User user = userService.findByUserName(userName);
+            boolean removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                journalEntryRepository.deleteById(id);
+                userService.saveUser(user);
+            }
+        }
+        catch (Exception e) {
+            throw new RuntimeException("An error occurred while deleting the entry.", e);
+        }
     }
 }
